moved getpass() method into a cross-plateform file rskbdinput.h/cc
