+++ libretroshare/src/util/rskbdinput.h
@@ -1,0 +1,25 @@
+/*******************************************************************************
+ * libretroshare/src/retroshare/util/rskbdinput.h                              *
+ *                                                                             *
+ * Copyright (C) 2019  Cyril Soler <csoler@users.sourceforge.net>              *
+ *                                                                             *
+ * This program is free software: you can redistribute it and/or modify        *
+ * it under the terms of the GNU Lesser General Public License as              *
+ * published by the Free Software Foundation, either version 3 of the          *
+ * License, or (at your option) any later version.                             *
+ *                                                                             *
+ * This program is distributed in the hope that it will be useful,             *
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of              *
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the                *
+ * GNU Lesser General Public License for more details.                         *
+ *                                                                             *
+ * You should have received a copy of the GNU Lesser General Public License    *
+ * along with this program. If not, see <https://www.gnu.org/licenses/>.       *
+ *                                                                             *
+ *******************************************************************************/
+
+namespace RsUtil {
+
+	std::string rs_getpass(const std::string& prompt,bool no_echo=true) ;
+
+}
